fix: renamed pagination component for bookings history
